Split out validator tests into method and model tests
